requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
